extract AttributionView
Not every page needs the getCustodian() method which allows us to show
the "provided by" text to give attribution to the custodian of a
register.

This extracts an AttributionView class which owns the custodian, and
ensures that we only extend it from classes which need the attribution
details; other classes still extend ThymeleafView directly.
