+++ src/main/java/uk/gov/register/presentation/view/ListVersionView.java
@@ -4,11 +4,11 @@ import com.fasterxml.jackson.annotation.JsonValue;
 import uk.gov.register.presentation.Version;
 import uk.gov.register.presentation.config.PublicBodiesConfiguration;
 import uk.gov.register.presentation.resource.RequestContext;
-import uk.gov.register.thymeleaf.ThymeleafView;
+import uk.gov.register.thymeleaf.AttributionView;
 
 import java.util.List;
 
-public class ListVersionView extends ThymeleafView {
+public class ListVersionView extends AttributionView {
     private final List<Version> versions;
 
     public ListVersionView(RequestContext requestContext, PublicBodiesConfiguration publicBodiesConfiguration, List<Version> versions) {
