+++ src/main/java/uk/gov/register/thymeleaf/BadRequestExceptionView.java
@@ -1,6 +1,5 @@
 package uk.gov.register.thymeleaf;
 
-import uk.gov.register.presentation.config.PublicBodiesConfiguration;
 import uk.gov.register.presentation.resource.RequestContext;
 
 import javax.ws.rs.BadRequestException;
@@ -8,8 +7,8 @@ import javax.ws.rs.BadRequestException;
 public class BadRequestExceptionView extends ThymeleafView {
     private final BadRequestException exception;
 
-    public BadRequestExceptionView(RequestContext requestContext, PublicBodiesConfiguration publicBodiesConfiguration, BadRequestException exception) {
-        super(requestContext, publicBodiesConfiguration, "400.html");
+    public BadRequestExceptionView(RequestContext requestContext, BadRequestException exception) {
+        super(requestContext, "400.html");
         this.exception = exception;
     }
