+++ src/main/java/uk/gov/register/thymeleaf/ThymeleafView.java
@@ -2,8 +2,6 @@ package uk.gov.register.thymeleaf;
 
 import io.dropwizard.views.View;
 import org.apache.commons.lang3.StringUtils;
-import uk.gov.register.presentation.config.PublicBodiesConfiguration;
-import uk.gov.register.presentation.config.PublicBody;
 import uk.gov.register.presentation.config.Register;
 import uk.gov.register.presentation.resource.RequestContext;
 
@@ -13,16 +11,12 @@ import javax.servlet.http.HttpServletResponse;
 import java.nio.charset.StandardCharsets;
 
 public class ThymeleafView extends View {
-
-    protected final PublicBodiesConfiguration publicBodiesConfiguration;
-    private final RequestContext requestContext;
-
+    protected final RequestContext requestContext;
     private String thymeleafTemplateName;
 
-    public ThymeleafView(RequestContext requestContext, PublicBodiesConfiguration publicBodiesConfiguration, String templateName) {
+    public ThymeleafView(RequestContext requestContext, String templateName) {
         super(templateName, StandardCharsets.UTF_8);
         this.requestContext = requestContext;
-        this.publicBodiesConfiguration = publicBodiesConfiguration;
     }
 
     @Override
@@ -67,10 +61,4 @@ public class ThymeleafView extends View {
     public HttpServletRequest getHttpServletRequest() {
         return requestContext.getHttpServletRequest();
     }
-
-    @SuppressWarnings("unused, used by templates")
-    public PublicBody getCustodian() {
-        return publicBodiesConfiguration.getPublicBody(getRegister().getRegistry());
-    }
 }
-
